refactor: refactored parts of code to better feat pushes from teammates.
