+++ src/components/Header/StyledHeader.ts
@@ -1,4 +1,4 @@
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 
 export const StyledHeader = styled.header`
     display: flex;
@@ -29,9 +29,7 @@ export const StyledHeader = styled.header`
         margin-right: 5%;
     }
 
-    button {
-        width: auto;
-    }
+   
 
     .menu {
         display: flex;
