+++ src/components/editForm/editForm.tsx
@@ -7,8 +7,6 @@ import { Controller, useForm } from "react-hook-form";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { iEditRequest } from "../../types/types";
-import { useContext } from "react";
-import { userContext } from "../../contexts/userContext/userContext";
 import { Input } from "../input/input";
 import { PulseLoader } from "react-spinners";
 import {} from "@mui/icons-material";
