+++ src/contexts/userContext/userContext.tsx
@@ -12,7 +12,6 @@ import { API } from "../../services/axios";
 
 import { AxiosError } from "axios";
 import { Toast } from "../../components/Toast";
-import { loadingContext } from "../loadingContext/loadingContext";
 import { useLoading } from '../../hooks/useLoading';
 
 interface iLoginError {
