+++ src/pages/ProfilePage.tsx
@@ -1,8 +1,8 @@
-import EditPage from "../components/edit/edit";
+
 import { Header } from "../components/Header/Header";
 import Modal from "../components/modal/modal";
-import { useModal } from "../contexts/modalContext/modalContext";
-
+import EditForm from '../components/editForm/editForm';
+import { useModal } from '../hooks/useModal';
 export function ProfilePage () {
     const { stateModal, showModal} = useModal()
 
@@ -10,7 +10,7 @@ export function ProfilePage () {
         <>
             <Header/>
             {stateModal && <Modal/>}
-            <button onClick={() => showModal(<EditPage/>)}>Abrir Modal de Editar</button>
+            <button onClick={() => showModal(<EditForm/>)}>Abrir Modal de Editar</button>
         </>
     )
 }
